Adding CAP to passive train_dataset
