+++ examples/train.ini
@@ -34,7 +34,7 @@ WEIGHT_DECAY_RATE = 0.00078125
 OPTIMIZER = MOM
 
 [self.dataset]
-DATASET_NAME = cifar10
+DATASET_NAME = passive_cifar10
 DATASET_DIR = ${vars:DATASET_DIR}
 TRAIN_SET_SIZE = 50000
 VALIDATION_SET_SIZE = 10000
@@ -55,7 +55,7 @@ DRIFT_X = 4
 DRIFT_Y = 4
 
 [self.train.train_control]
-TRAINER = passive
+TRAINER = simple
 TRAIN_BATCH_SIZE = 200
 EVAL_BATCH_SIZE  = 2200
 ROOT_DIR = ${vars:ROOT_DIR}
