+++ lib/datasets/active_dataset.py
@@ -1,27 +1,22 @@
 import numpy as np
-from lib.datasets.dataset import DataSet
+from lib.datasets.passive_dataset import PassiveDataSet
 
-class ActiveDataSet(DataSet):
+class ActiveDataSet(PassiveDataSet):
     def __init__(self, *args, **kwargs):
-        super(DataSet, self).__init__(*args, **kwargs)
+        super(ActiveDataSet, self).__init__(*args, **kwargs)
 
         self.n_clusters = self.prm.dataset.N_CLUSTERS
-        self.cap        = self.prm.dataset.CAP
-
-        self.pool = []  # starting from zero indices in the pool
-        self.available_samples = range(self.size)
 
         self.initialize_pool()
 
     def print_stats(self):
         super(ActiveDataSet, self).print_stats()
         self.log.info(self.__str__() + ': N_CLUSTERS: {}'.format(self.n_clusters))
-        self.log.info(self.__str__() + ': CAP: {}'.format(self.cap))
 
     def initialize_pool(self):
-        while len(self.pool) < self.batch_size:
-            self.log.info('Small pool length: {}. Adding to pool {} random elements'.format(len(self.pool), self.n_clusters))
-            self.update_pool_rand()
+        while self.pool_size() < self.batch_size:
+            self.log.info('Small pool size: {}. Adding to pool {} random elements'.format(self.pool_size(), self.n_clusters))
+            self.update_pool()
 
     def update_pool(self, n_clusters=None, indices=None):
         if indices is not None:
@@ -36,6 +31,10 @@ class ActiveDataSet(DataSet):
         self.update_pool_with_indices(indices)
 
     def update_pool_with_indices(self, indices):
+        self.assert_unique_indices(indices)  # time consuming. There is no sense checking it for passive dataset.
+        super(ActiveDataSet, self).update_pool_with_indices(indices)
+
+    def assert_unique_indices(self, indices):
         for index in indices:
             if index in self.pool:
                 err_str = 'update_pool_with_indices: index {} is already in pool.'.format(index)
@@ -44,7 +43,4 @@ class ActiveDataSet(DataSet):
             if 'err_str' in locals():
                 self.log.error(err_str)
                 raise AssertionError(err_str)
-        self.pool += indices
-        self.pool = sorted(self.pool)
-        self.available_samples = [i for j, i in enumerate(self.available_samples) if i not in self.pool]
-        self.log.info('updating pool length to {}'.format(len(self.pool)))
+
